+++ NEWS.md
@@ -1,9 +1,19 @@
+Andromeda 1.0.0
+===============
+
+Changes
+
+- Switch backend from SQLite to DuckDb for greater performance in terms of speed and disk space.
+
+- Added `andromedaMemoryLimit` option to control the maximum amount of memory Andromeda is allowed to use (in GB).
+
+
 Andromeda 0.6.7
 ===============
 
 Bugfixes
 
-- Fix `isAndromedaTable()` when table belongs to descendant of Andromda.
+- Fix `isAndromedaTable()` when table belongs to descendant of Andromeda.
 
 
 Andromeda 0.6.6